Move transform API to struct methods for easier memory management
When combining transformations, previously it was necessary to release
every intermediate state manually. This is complex and error-prone. By
using struct methods, the image can release its previous memory itself.

The package-based API is still kinda nice for a number of reasons, so
we may bring it back someday—but for now it's easier to maintain one.

Signed-off-by: Evan Owen <evan@zinc.it>